- Changes Google Cluster Data readers' job_id, task_index and machine_id   fields to a 64 bits integer (Java's long type),   to conform to the trace files data format. - Changes Entities' ID from int to long,   enabling large scale simulations created from workload files   (such as Google Cluster Data) when there is   a high number of objects (such as physical machines   and jobs) that exceeds the int range. - Changes calls such as Integer.valueOf to Integer.parseInt to avoid   unnecessary boxing. - Defines default values for number fields in Google trace files as:   (i) -1 for important fields such as machine id (meaning there is no machine allocated)   or maximum values (meaning there is no limit);   (ii) 0 for means and minimum values.

+++ cloudsim-plus-examples/src/main/java/org/cloudsimplus/examples/brokers/CloudletToVmMappingBestFit.java
@@ -115,7 +115,7 @@ public class CloudletToVmMappingBestFit {
         simulation.start();
 
         final List<Cloudlet> finishedCloudlets = broker0.getCloudletFinishedList();
-        finishedCloudlets.sort(Comparator.comparingInt(Cloudlet::getId));
+        finishedCloudlets.sort(Comparator.comparingLong(Cloudlet::getId));
         new CloudletsTableBuilder(finishedCloudlets).build();
     }
