+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/hosts/HostNull.java
@@ -58,7 +58,7 @@ final class HostNull implements Host {
     @Override public Datacenter getDatacenter() {
         return Datacenter.NULL;
     }
-    @Override public int getId() {
+    @Override public long getId() {
         return -1;
     }
     @Override public double getMaxAvailableMips() {
@@ -148,7 +148,7 @@ final class HostNull implements Host {
     @Override public String toString() {
         return "Host.NULL";
     }
-    @Override public void setId(int id) {/**/}
+    @Override public void setId(long id) {/**/}
     @Override public double getTotalMipsCapacity() { return 0.0; }
     @Override public long getNumberOfFailedPes() { return 0; }
     @Override public List<Pe> getWorkingPeList() { return Collections.emptyList(); }
